removed redundant methods, renamed graphobject class,option to switch all/cation-anion bond
